aws-cdk-deployment/requirements.txt
-e .
aws_cdk.aws_s3
aws_cdk.aws_ec2
aws_cdk.aws_ecs
aws_cdk.aws_ecr
